services/surya/requirements.txt
# Surya OCR Service Requirements
# Fixed at transformers 4.36.2 for Surya compatibility

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.5.3,<3.0.0

# Surya OCR dependencies
surya-ocr
transformers==4.36.2
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.4.0

# Additional utilities
numpy>=1.24.0